+++ ChangeLog.md
@@ -7,3 +7,8 @@
 - Make the readme for the project.
 - Clone the template for the plugin.
 - Make the manifest for the project.
+
+### 2023/03/06 Version 1.0
+
+- The plugin is now functional.
+- Possible to change the library of medias.